Added and *tested* openmp directives to the Turbulence and TurbGOTM modules. The Outwatch logs evidence performance gains in a dual-core machine between 10% to up to 40% in individual subroutines. However, the overall performance is diminished between 2% to 12%, basically because of the overhead introduced by the parallelization.
Perhaps with a higher number of cores (>2) overall performance will improve. Perhaps when *all* the code is parallelized with openmp directives, the gain in individual subroutines will exceed the loss introduced in the overhead.

Did a report that I uploaded to \\einstein\parallel on that subject. Also edited the common xlsx document.

git-svn-id: https://mohid.svn.codeplex.com/svn@41767 81fbe566-5dc4-48c1-bdea-7421811ca204