1. Added the external lib PhreeqCLIB 2. Added the modules ModulePhreeqC and ModulePhreeqCData 3. Updated the files ModulePorousMediaProperties, ModuleInterface and ModuleGlobalData to be able to use PhreeqCLIB 4. Updated MOHIDBase1.vfproj to include ModulePhreeqC and ModulePhreeqCData
Observations about the project.
To use PhreeqC must be defined (preprocessor definitions) the keywords _PHREEQC_ and _LONG_LINE_LENGTH.
Default Real KIND must be set to 8 (/real_size:64)

IMPORTANT:
The code is not fully working yet.
Need some corrections and has a problem with Units conversion that is being work out with David L. Parkhurst (one of the authors of the PhreeqC). As soon as this be done, a new working version will be commited.

git-svn-id: https://mohid.svn.codeplex.com/svn@42952 81fbe566-5dc4-48c1-bdea-7421811ca204