The modules ModulePhreeqC and ModulePhreeqCData were removed from the solution.
This was done due to the fact that these modules were programmed to work only in double precision, leading to compiling time errors when using single precision.

Once a satisfactory solution is found, these modules will be part of the project again.

They can still be included in projects that use double precision.

git-svn-id: https://mohid.svn.codeplex.com/svn@43217 81fbe566-5dc4-48c1-bdea-7421811ca204