Removed the compilation option "-heap-arrays 64" from both the linux and the windows default configuration solutions in MOHIDWater, MohidBase1 and MohidBase2, since it induced a memory leak that would eventually blow-up simulations with an "insufficient virtual memory" error.
This was detected independently in Linux (by A.V.) and in Windows (by F.C.). This anomalous memory leak was reported in this intel forum thread:

http://software.intel.com/en-us/forums/showthread.php?t=64247

...but it is thought to be resolved as of ifort 11.1. However, both linux and windows test-cases were using ifort 11.1 or greater. Could it be that the memory leak with the -heap-array option wasn't completely fixed in ifort 11.1?

Another way that is reported in the above thread to solve the memory leak problem when using -heap-arrays is to replace POINTER with ALLOCATABLE in all the code ... Maybe this could work.

In the meantime, MOHID default compilation options are set without -heap-arrays ...

git-svn-id: https://mohid.svn.codeplex.com/svn@67018 81fbe566-5dc4-48c1-bdea-7421811ca204