Openmp directives implemented and tested in module Hydrodynamic, in subroutine Velocity_VerticalAdvection.
A bug was found that didn't initialize the variable NearBoundary to .false., maintaining it at .true. once it was set. This bug was the reason why results with openmp for different number of threads were different (when testing the openmp directives on it).

This also meant that, during most of the simulation time, the vertical advection method (specified by the keyword ADV_METHOD_V) was automatically overridden to First-Order Upwind, thus disabling any higher-order method.

The bug is fixed, but this also means that every project with keyword ADV_METHOD_V set to a value other than 1 (typically the  choice of 4 is given in many projects) may yield a different behavior from the user point of view.

Users with projects already underway should comment the ADV_METHOD_V keyword from their hydrodynamic configuration files, to keep similar results than those in the past.

A higher-order numerical method for vertical advection increases the precision of the calculations (ie reduces the error) BUT it also increases the chances of instabilities to occur in the model.

In hydrostatic, boussinesq-approximated models, higher-order numerical methods are important for horizontal advection but are less relevant for vertical advection. Indeed,  along the vertical, mixing processes prevail over advection; hence, a first-order upwind method is a perfectly reasonable choice.

git-svn-id: https://mohid.svn.codeplex.com/svn@69195 81fbe566-5dc4-48c1-bdea-7421811ca204