ConvertToHDF5: Correction of a BUG that prevented the compilation without NetCDF
ModuleBasin: 
Changed the routine PredictNewDT to correct a problem of convergence on PorousMedia when the NExtDT is to big and there is a rain event at the end of it.

ModulePorousMedia:
The convergence test in routine "variation_test" now is coherent with RunOff and DrainageNetwork. The keyword THETA_TOLERANCE is now deprecated and instead the keyword STABILIZE_FACTOR must be used, with a value between 0 and 1 (decimal % of the old volume that is allowed). 


git-svn-id: https://mohid.svn.codeplex.com/svn@86297 81fbe566-5dc4-48c1-bdea-7421811ca204