ModuleBenthicEcology: Commented an unused variable and split a line to be less than 132 characters long.
ModuleDischarges: Initialization of variables

ModuleGlobalData: Initialization of variables and split a line to be less than 132 characters long.

ModuleInterface: Initialization of variables

ModuleMacroAlgae: Split a line to be less than 132 characters long.

ModuleTask2000: Change of two Real*8 to Real(8), that is the standard.

ModuleWaterQuality: Initialization of variables

ModuleAdvectionDiffusion: Change of the operators == and /= to .eq. and .ne.

ModuleAtmosphere: Initialization of variables

ModuleBoxDiff: Initialization of variables

ModuleFillMatrix: Initialization of Variables and the ieee_arithmetic instructions are now under conditional compilation (but enabled by default)

ModuleGeometry: Initialization of variables

ModuleHorizontalGrid: Initialization of variables

ModuleStatistic: Initialization of variables (INCOMPLETE)

ModuleFreeVerticalMovement: Initialization of variables and correction of a BUG in the Kill_FreeVerticalMovement subroutine

ModuleGOTM: No changes (???)

ModuleHydrodynamic: Initialization of variables and a variable unused was commented in subroutine BenthicEcology_Processes

ModuleInterfaceWaterAir: Initialization of variables

ModuleLagrangian: Spaces

ModuleModel: Initialization of variables

ModuleWaterProperties: Split a line to be less than 132 characters long.

Makefiles: Update to include new modules.

git-svn-id: https://mohid.svn.codeplex.com/svn@86763 81fbe566-5dc4-48c1-bdea-7421811ca204