BUG FIX.
Fix bug in ModulePorousMediaProperties, subroutine ComputeVolumes, where the K was being declared as PRIVATE in the second OpenMP looping, but it's value was not being changed inside the looping. In this case, the K must be declared SHARED (it's the default) or FIRSTPRIVATE, so each copy of the variable will be initialized before the thread starts the looping. Declaring as "PRIVATE" was causing it's value to be "undetermined". In this case, the SHARED (default) was the choice as there is no need to every thread to have it's own copy of the variable.

git-svn-id: https://mohid.svn.codeplex.com/svn@89027 81fbe566-5dc4-48c1-bdea-7421811ca204