This commit add features and change the behaviour of some functions related with DYNAMIC PROPERTIES. These changes allow the use of "dynamic properties" by the modules in a more transparent way, requiring much less coding (or none, in some cases), and will make new developments with recourse to dynamic properties simpler.
ModuleGlobalData:
- Add the 'CheckDynamicPropertyName' function, with the same behaviour of the CheckPropertyName function, but applied to the list of dynamic properties.
- Change the CheckPropertyName function. Now, if the property name is not found in the property list, the routine also looks in the dynamic property list for it.
- Change the RegisterDynamicProperty function. 
Now:
a) The routine looks into the global data property list to see if the property name already exists there. If it already exists there, the model stops with a message.
b) If the property already is registered in the dynamic property list, the id of the property (already registered) is returned and the property is not registered again.
c) It is possible to know if the property was already registered before the call to the function through the optional PropertyExists argument. If the property already exists in the dynamic list, this will be set to .TRUE., or .FALSE. otherwise.
- Change the following routines to also look into the dynamic property list if the property is not found in the global data property list: GetPropertyName and GetPropertyIDNumber

ModuleFunctions:
- Change the ConstructPropertyID function. Now, after getting the property name (NAME), it looks for the DYNAMIC keyword (1 or 0). If it exists and is set to 1, the property is considered as a Dynamic Property, and the routine RegisterDynamicProperty is called to register the property in the dynamic list before continuing the execution of the code.

git-svn-id: https://mohid.svn.codeplex.com/svn@89515 81fbe566-5dc4-48c1-bdea-7421811ca204