MOHID Water & MOHID Land - added the option not to stop the model when bathymetry checks are done (because of geometry and isolated cells checks) since the updated bathymetry is already in model memory. The keyword added is STOP_ON_BATHYMETRY_CHANGE : 0 in Model_X.dat. By default the value is 1 and so the model stops after bathymetry change and asks the user to change nomfich (the same behavior from years now). If the user explicitly puts value 0 than the new changed bathymetry is still saved to new file and warnings still exists but model continues using the updated bathymetry.
MOHID Land - added the option to run the model without removing depressions in DTM and using drainage network so that in detailed areas where DTM is known dont need to apply remove depressions that usually make the DTM strange (in flat areas) or we do not want to remove REAL depressions. The keyword is SET_RIVER_POINTS_FROM_DN : 1 in BasinGeometry_X.dat. By default the value is 0 where it behaves as before. If the user explicitly puts value 1 the river points will not be computed as before from DTM but setted from Drainage Network file. So that drainage Network and DTM are consistent (e.g. runoff and DN levels), in this particular case the nodes are verified if node TERRAIN_LEVEL is consistent with DTM. if not, node TERRAIN_LEVEL and node HEIGHT are adjusted to DTM elevations.
A typical use of this would be to generate Drainage Network from removed depressions DTM but use original DTM to run (that describes REAL depressions).

git-svn-id: https://mohid.svn.codeplex.com/svn@89588 81fbe566-5dc4-48c1-bdea-7421811ca204