modulesediment - After validation with experiments data in channels and tests in schematic cases and real cases to verify all module features (without the waves action), this can be considered the first stable version of Module Sediment.
modulefreeverticalmovement - A maximum value was imposed to the settling velocity of micro and macroflocs (Ws_Type :6).

moduleinterfacesedimentwater - A method to compute the rugosity field depending on sediment granulometry and hydrodynamic conditions was implemented (based on van Rijn, 2007). For now, it works only when running module sediment. 

moduleturbgotm - Values of Tke, L, and eps at open boundary points were set equal to the values of the nearest interior point (Null_gradient). This avoid discontinuities at the boundary and problems when running with Ws_Type :6 in modulefreeverticalmovement, which needs eps results.

git-svn-id: https://mohid.svn.codeplex.com/svn@89647 81fbe566-5dc4-48c1-bdea-7421811ca204