Error correct in the slip condition (land boundary)
An error detected by RivageProtech related with the slip  condition was
corrected.
First step in breaking down the MoveParticHorizontal subrotine. The idea
is to breakdown this subroutine form the present 1000 code lines to
subroutines in average with 100 code lines. Due to the complexity of the
data flux this must be done carefully. The first step was to create a
subroutine to deal with boom and thin walls.