Correction of the direction of the bedload component normal to the currents direction
The normal component of the bedload to the currents direction is caused
due to waves asymetry and should be shoreward. Schematic tests showed
that this component was in the wrong direction. This bug was fixed by
simply considering CWphi = WaveAngle - CurrentAngle, instead of
CurrentAngle - WaveAngle. CWphi is used in subroutine
CurrentPlusWavesBedload (module Sediment).