Change the autmatic domain decomposition procedure
Change the autmatic domain decomposition procedure. In the past the
decomposition was done along the lines. Now check if the grid lines are
or not more than the columns. If the colums are more than the automatic
decomposition is along the columns. Now if the manual mapping input file
of the domain decomposition does not exist a warnning is written in the
screen.