Important error related with MPI
For the MPI version of MOHID there was an error for the follow
conditions:
- nested model with a geometry different from the father. In this
conditions the values imposed and relax in the boundary was always equal
to the initial condition. Code cleanning. Now if noty SIMPLE_OUTPUT the
field used in nudging process is written in the hdf5 output file in the
group Results\Assimila