Allow to write in a output (hdf5) of the interpolated tidal components
Allow to write in a output (hdf5) of the interpolated tidal components.
The only needs to define in the file (define in nomfich.dat with the
keyword IN_TIDES) add the follow keyword with the name of the output
file FILE_TIDAL_COMPONENTS_OUTPUT :
..\res\TidalComponents_database_Fes2012_Portugal.hdf5
This also output the extra tidal components estimated based in others
(admittance option).