Critical error related with hotstart solved
Critical error related with hotstart solved. The last update assumed an error when a hdf5 file was read using binary fortran read instructions. In practise this error is not detected by fortran. The model is able to read from the hdf5 wrong values. Now the model is able to detect the format of the initial results of the hydrodynmaic or turbulence modules (bin or hdf5);
Important changes were also added related with imposing in the boundary a wind wave and the absorption of wind waves by land cells.