Changed convert2netcdfName and operations order
Name of the convert2netcdf was different from the module name, giving an error in linux.
In functions, the order of a calculation of a function invoqued by an optimized routine was changed because it generated very different results over time. It is now using the same order of operation as the non-optimized version.