Flow scale factor & Geo 2 cartesian exact
keyword FLOW_SCALE_FACTOR can be use in the discharges and lagrangianglobal modules. This allow the definition of a generic flow time serie that is multiply by value define via this keyword.
Now the convertion from geo to cartesian and vice-versa in the lagrangian global is independent of the grid being used.