Bug in the wind drag coefficient computation
Bug in the  wind drag coefficient computation based on Large & Pond, 1981 considering the modifications for low wind speeds proposed by Trenberth et al. (1990)
Error was focus in the wind velocities between 1 a 3 m/s. For 1 m/s the error was 0 but for 3 m/s the drag coefficient was 4.7 times higher than the correct value. In the 3 m/s limit wind forcing should  generate some noise because the drag coeffcient is oscilatting along this limit between 1.14e-3 and 5.3e-3

If the keyword DEFINE_CDWIND is not define the Large & Pond (1981) formulation is the default one.