remodeling of readoptions for readability
Created several smaller routines to replace the readoptions routine in fillmatrix. old version was not readable and creates the idea that it only reads user options (but it does everything). Also added a check in assimilation for new upscaling domain, in order to allocate new grid modules when a new domain is used. (old version creates new modules for every new property and in upscaling one can recieve many with the same name but on different domains)