Output of the phase delay for each grid cell relative to reference gauge.
Output of the phase delay relative to reference gauge. The new output is hydrodynamic_x_HighLowTide.hdf5. They are 2D results. The iinstant are of the low and high tide of the reference tidal gauge. Aditionally the high and low tide reach in each cell is also in the output.