Relative position in cell using IWD
Previous version the relative position of a point in a cell was done using a bi-linear approach. It was to complex and the method implemented was not generic enough for curvilin ear cells. The new method is based in the IWD (inverse weight distance) interpolation method. It ios more simple to code and more robust.