Correction of continuous calculation errors in vegetation. Added Kc evolution. Changes in reservoirs module.
Continuous calculation errors in vegetation: with continuous calculation the module was called twice in a day causing duplication of vegetation growth.
Kc evolution: the Kc can be now calculated according to FAO56 as a function of LAI.
Changes in reservoirs module: the option of having a maximum volume each month and to introduce the year of reservoir construction was added.